rename unit tests for actionQueue and filter utility functions
